Move common utilities into new pkg
