+++ common/ascii.go
@@ -1,4 +1,4 @@
-package client
+package common
 
 var ASCII = []string{"    d8b",
 	"     ?88",
