+++ common/writer.go
